Implement houses module
- Implement house module in admin routes;
- Create vehicles routes;
