+++ app/views/admin/houses/houses/show.slim
