+++ app/views/admin/houses/residents/index.slim
