+++ app/views/admin/houses/residents/show.slim
